link tag: add link tags into image to give necessary info
when use loader to verify the image security, we need to
know the image length. But with a XIP mode, we can not simply
append these info to the header. Since the vectors are always at
the beggining of the image, we can insert these info after vectors.
For unicorn, we specify 0x400 for the offset of these info at link
phase automatically.

Change-Id: I5ba3c123e7947ccb2d09636f1094ff26e3cd4f06
Signed-off-by: QianWenfa <qianwenfa@xiaomi.com>